+++ lib/main.py
@@ -11,7 +11,7 @@ from lib import log
 if TYPE_CHECKING:  # pragma: no cover
     from config import Config
 
-__version__ = "v4.0"
+__version__ = "v4.1"
 
 __doc__ = """
 Schedule a check-in: