+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 When upgrading to a new version, make sure to follow the directions under the "Upgrading" header of the corresponding version.
 If there is no "Upgrading" header for that version, no post-upgrade actions need to be performed.
 
+## Upcoming
+
+### Bug Fixes
+- Handle cases where no fares are available for a flight
+([#86](https://github.com/jdholtz/auto-southwest-check-in/issues/86))
+
+
 ## 4.0 (2023-05-06)
 
 ### New features